feat: restructure application entry point and update Dockerfile
- Moved the application entry point from `src/index.js` to `server.js` for better organization and clarity.
- Updated the `Dockerfile` to reflect the new entry point by changing the CMD instruction to start the application from `server.js`.
- Modified the `package.json` script for development to use `server.js` instead of `src/index.js`.
- Created a new `src/app.js` file that initializes the Express application, sets up middleware, and defines routes, enhancing modularity and maintainability.
