+++ src/app.js
@@ -10,7 +10,6 @@ const userRouter = require('./routes/user.routes.js');
 const authRouter = require('./routes/auth.routes.js');
 
 const app = express();
-const PORT = 3001;
 const NODE_ENV = process.env.NODE_ENV || 'development';
 
 dotenv.config();
@@ -35,14 +34,5 @@ app.get('/', (req, res) => {
   res.status(200).json({ message: 'Hello from Nodesocial API' });
 });
 
-const server = app.listen(PORT, () => {
-  console.log(`App running in mode at http://127.0.0.1:${PORT}`);
-});
-
-process.on('unhandledRejection', (err, promise) => {
-  console.log(`Error: ${err.message}`.red);
-
-  server.close(() => process.exit(1));
-});
 
 module.exports = app;
